python/requirements.txt
python-dotenv
neo4j-driver==5.10.0
msal